+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [tanjun.abc.CLient.load_directory][] and [tanjun.abc.Client.load_directory_async][] for loading all
+  the modules in a directory.
+- `path` properties to [tanjun.errors.FailedModuleLoad][] and [tanjun.errors.FailedModuleUnload][]
+  to indicate which module failed to load/unload.
+- [tanjun.errors.FailedModuleImport][] specialisation of [tanjun.errors.FailedModuleLoad][] for
+  differentiating between the two cases when necessary.
 
 ## [2.6.1a1] - 2022-08-29
 ### Added